flip_x - wip t15 vcol flipping, scenery gaps bug fix; switched to proper stdint types

+++ src/build_merge_matrix_random.c
@@ -6,30 +6,30 @@
 typedef struct matrix_merge_thread_input_struct
 {
     ENTRY *elist;
-    int entry_count;
-    unsigned int *best_zone_ptr;
-    long long int *best_max_ptr;
+    int32_t entry_count;
+    uint32_t *best_zone_ptr;
+    int64_t *best_max_ptr;
     ENTRY *best_elist;
     LIST entrs;
-    int **entry_mtrx;
-    int *conf;
-    int max_pay;
-    int iter_cnt;
-    int *curr_iter_ptr;
+    int32_t **entry_mtrx;
+    int32_t *conf;
+    int32_t max_pay;
+    int32_t iter_cnt;
+    int32_t *curr_iter_ptr;
     double mlt;
-    int rnd_seed;
-    int thread_idx;
-    int *running_threads;
+    int32_t rnd_seed;
+    int32_t thread_idx;
+    int32_t *running_threads;
     pthread_mutex_t *mutex_running_thr_cnt;
     pthread_mutex_t *mutex_best;
     pthread_mutex_t *mutex_iter;
-    int chunk_border_sounds;
+    int32_t chunk_border_sounds;
     MATRIX_STORED_LLS stored_lls;
 } MTRX_THRD_IN_STR;
 #endif // COMPILE_WITH_THREADS
 
 // asking user parameters for the method
-void ask_params_matrix(double *mult, int *iter_count, int *seed, int *max_payload_limit)
+void ask_params_matrix(double *mult, int32_t *iter_count, int32_t *seed, int32_t *max_payload_limit)
 {
     printf("\nMax payload limit? (usually 21 or 20)\n");
     scanf("%d", max_payload_limit);
@@ -78,17 +78,17 @@ void *build_matrix_merge_random_util(void *args)
 {
     char temp[6] = "";
     char temp2[6] = "";
-    int limit_reached = 0, best_reached = 0;
+    int32_t limit_reached = 0, best_reached = 0;
 
     MTRX_THRD_IN_STR inp_args = *(MTRX_THRD_IN_STR *)args;
     ENTRY clone_elist[ELIST_DEFAULT_SIZE];
-    int entry_count = inp_args.entry_count;
+    int32_t entry_count = inp_args.entry_count;
     RELATIONS array_repr_untouched = build_transform_matrix(inp_args.entrs, inp_args.entry_mtrx, inp_args.conf, inp_args.elist, entry_count);
     RELATIONS array_representation = build_transform_matrix(inp_args.entrs, inp_args.entry_mtrx, inp_args.conf, inp_args.elist, entry_count);
     srand(inp_args.rnd_seed);
-    int iter_count = inp_args.iter_cnt;
-    int curr_i;
-    int t_id = inp_args.thread_idx;
+    int32_t iter_count = inp_args.iter_cnt;
+    int32_t curr_i;
+    int32_t t_id = inp_args.thread_idx;
     double mult = inp_args.mlt;
 
     while (1)
@@ -104,12 +104,12 @@ void *build_matrix_merge_random_util(void *args)
             break;
 
         // restore stuff
-        for (int j = 0; j < entry_count; j++)
+        for (int32_t j = 0; j < entry_count; j++)
             clone_elist[j] = inp_args.elist[j];
         memcpy(array_representation.relations, array_repr_untouched.relations, array_representation.count * sizeof(RELATION));
 
         // second half of the matrix merge slightly randomised and ran on clone elist
-        for (int j = 0; j < array_representation.count; j++)
+        for (int32_t j = 0; j < array_representation.count; j++)
             array_representation.relations[j].value = array_repr_untouched.relations[j].value * randfrom(1.0, mult);
         qsort(array_representation.relations, array_representation.count, sizeof(RELATION), relations_cmp);
 
@@ -121,21 +121,21 @@ void *build_matrix_merge_random_util(void *args)
         qsort(payloads.arr, payloads.count, sizeof(PAYLOAD), cmp_func_payload);
 
         // how many cam paths' payload it takes into consideration, max 8 but if the level is smaller it does less, also it could do 9 too but 8 is a nicer number
-        int src_depth = min(payloads.count, 8);
+        int32_t src_depth = min(payloads.count, 8);
 
         // gets current 'score', its multiplied by 100 instead of being bit shifted each iter so its legible in the console
-        long long int curr = 0;
-        for (int j = 0; j < src_depth; j++)
+        int64_t curr = 0;
+        for (int32_t j = 0; j < src_depth; j++)
         {
-            curr = curr * (long long int)100;
-            curr += (long long int)payloads.arr[j].count;
+            curr = curr * (int64_t)100;
+            curr += (int64_t)payloads.arr[j].count;
         }
         pthread_mutex_lock(inp_args.mutex_best);
-        int is_new_best = 0;
+        int32_t is_new_best = 0;
         // if its better than the previous best it gets stored in best_elist and score and zone are remembered
         if (curr < *inp_args.best_max_ptr)
         {
-            for (int j = 0; j < entry_count; j++)
+            for (int32_t j = 0; j < entry_count; j++)
                 inp_args.best_elist[j] = clone_elist[j];
             *inp_args.best_max_ptr = curr;
             *inp_args.best_zone_ptr = payloads.arr[0].zone;
@@ -145,8 +145,8 @@ void *build_matrix_merge_random_util(void *args)
         pthread_mutex_lock(inp_args.mutex_iter);
         curr_i = *inp_args.curr_iter_ptr;
 
-        long long int cr_max = *inp_args.best_max_ptr;
-        for (int j = 0; j < src_depth - 1; j++)
+        int64_t cr_max = *inp_args.best_max_ptr;
+        for (int32_t j = 0; j < src_depth - 1; j++)
             cr_max = cr_max / 100;
 
         if (cr_max <= inp_args.max_pay)
@@ -190,25 +190,25 @@ void *build_matrix_merge_random_util(void *args)
 /** \brief
  *
  * \param elist ENTRY*
- * \param entry_count int
- * \param chunk_border_sounds int
- * \param chunk_count int*
- * \param config int*
+ * \param entry_count int32_t
+ * \param chunk_border_sounds int32_t
+ * \param chunk_count int32_t*
+ * \param config int32_t*
  * \param permaloaded LIST
  *
  */
-void build_matrix_merge_random_thr_main(ENTRY *elist, int entry_count, int chunk_border_sounds, int *chunk_count, int *config, LIST permaloaded)
+void build_matrix_merge_random_thr_main(ENTRY *elist, int32_t entry_count, int32_t chunk_border_sounds, int32_t *chunk_count, int32_t *config, LIST permaloaded)
 {
 
     // asking user parameters for the method
     double mult;
-    int iter_count, seed, max_payload_limit;
+    int32_t iter_count, seed, max_payload_limit;
     ask_params_matrix(&mult, &iter_count, &seed, &max_payload_limit);
 
     // thread count, mutexes
-    int t_count = 0;
-    int curr_iter = 0;
-    int running_threads = 0;
+    int32_t t_count = 0;
+    int32_t curr_iter = 0;
+    int32_t running_threads = 0;
     pthread_mutex_t running_mutex = PTHREAD_MUTEX_INITIALIZER;
     pthread_mutex_t iter_mutex = PTHREAD_MUTEX_INITIALIZER;
     pthread_mutex_t best_mutex = PTHREAD_MUTEX_INITIALIZER;
@@ -228,19 +228,19 @@ void build_matrix_merge_random_thr_main(ENTRY *elist, int entry_count, int chunk
     build_assign_primary_chunks_all(elist, entry_count, chunk_count);                           // chunks start off having one entry per chunk
 
     // get occurence matrix
-    int permaloaded_include_flag = config[CNFG_IDX_MTRX_PERMA_INC_FLAG];
+    int32_t permaloaded_include_flag = config[CNFG_IDX_MTRX_PERMA_INC_FLAG];
     LIST entries = build_get_normal_entry_list(elist, entry_count);
     if (permaloaded_include_flag == 0)
-        for (int i = 0; i < permaloaded.count; i++)
+        for (int32_t i = 0; i < permaloaded.count; i++)
             list_remove(&entries, permaloaded.eids[i]);
 
-    int **entry_matrix = build_get_occurence_matrix(elist, entry_count, entries, config);
+    int32_t **entry_matrix = build_get_occurence_matrix(elist, entry_count, entries, config);
 
     // clone elists that store the current iteration and the best iretation
     ENTRY best_elist[ELIST_DEFAULT_SIZE];
     // for keeping track of the best found
-    long long int best_max = 9223372036854775807; // max signed 64b int
-    unsigned int best_zone = 0;
+    int64_t best_max = 9223372036854775807; // max signed 64b int32_t
+    uint32_t best_zone = 0;
 
     MATRIX_STORED_LLS stored_lls = build_matrix_store_lls(elist, entry_count);
 
@@ -248,7 +248,7 @@ void build_matrix_merge_random_thr_main(ENTRY *elist, int entry_count, int chunk
     pthread_t *threads = (pthread_t *)malloc(t_count * sizeof(pthread_t));
     MTRX_THRD_IN_STR *thread_args = (MTRX_THRD_IN_STR *)malloc(t_count * sizeof(MTRX_THRD_IN_STR));
 
-    for (int i = 0; i < t_count; i++)
+    for (int32_t i = 0; i < t_count; i++)
     {
         thread_args[i].best_elist = best_elist;
         thread_args[i].best_max_ptr = &best_max;
@@ -275,7 +275,7 @@ void build_matrix_merge_random_thr_main(ENTRY *elist, int entry_count, int chunk
         running_threads++;
         pthread_mutex_unlock(&running_mutex);
 
-        int rc = pthread_create(&threads[i], NULL, build_matrix_merge_random_util, (void *)&thread_args[i]);
+        int32_t rc = pthread_create(&threads[i], NULL, build_matrix_merge_random_util, (void *)&thread_args[i]);
         if (rc)
             printf("Error while creating %d. thread - err %d\n", i, rc);
     }
@@ -287,7 +287,7 @@ void build_matrix_merge_random_thr_main(ENTRY *elist, int entry_count, int chunk
     }
 
     // copy in the best one
-    for (int i = 0; i < entry_count; i++)
+    for (int32_t i = 0; i < entry_count; i++)
     {
         elist[i] = best_elist[i];
         if (elist[i].chunk >= *chunk_count)
@@ -295,10 +295,10 @@ void build_matrix_merge_random_thr_main(ENTRY *elist, int entry_count, int chunk
     }
 
     // cleanup
-    for (int i = 0; i < stored_lls.count; i++)
+    for (int32_t i = 0; i < stored_lls.count; i++)
         free(stored_lls.stored_lls[i].full_load.eids);
     free(stored_lls.stored_lls);
-    for (int i = 0; i < entries.count; i++)
+    for (int32_t i = 0; i < entries.count; i++)
         free(entry_matrix[i]);
     free(entry_matrix);
     free(threads);
@@ -308,17 +308,17 @@ void build_matrix_merge_random_thr_main(ENTRY *elist, int entry_count, int chunk
 }
 #endif // COMPILE_WITH_THREADS
 
-MATRIX_STORED_LLS build_matrix_store_lls(ENTRY *elist, int entry_count)
+MATRIX_STORED_LLS build_matrix_store_lls(ENTRY *elist, int32_t entry_count)
 {
     MATRIX_STORED_LLS stored_stuff;
     stored_stuff.count = 0;
     stored_stuff.stored_lls = NULL;
 
-    int i, j, l, m;
+    int32_t i, j, l, m;
     for (i = 0; i < entry_count; i++)
         if (build_entry_type(elist[i]) == ENTRY_TYPE_ZONE && elist[i].data != NULL)
         {
-            int cam_count = build_get_cam_item_count(elist[i].data) / 3;
+            int32_t cam_count = build_get_cam_item_count(elist[i].data) / 3;
             for (j = 0; j < cam_count; j++)
             {
                 LOAD_LIST load_list = build_get_load_lists(elist[i].data, 2 + 3 * j);
@@ -342,7 +342,7 @@ MATRIX_STORED_LLS build_matrix_store_lls(ENTRY *elist, int entry_count)
 
                     if (list.count > 0)
                     {
-                        int stored_c = stored_stuff.count;
+                        int32_t stored_c = stored_stuff.count;
                         if (stored_c > 0)
                             stored_stuff.stored_lls = (MATRIX_STORED_LL *)realloc(stored_stuff.stored_lls, (stored_c + 1) * sizeof(MATRIX_STORED_LL));
                         else
@@ -363,12 +363,12 @@ MATRIX_STORED_LLS build_matrix_store_lls(ENTRY *elist, int entry_count)
     return stored_stuff;
 }
 
-PAYLOADS build_matrix_get_payload_ladder(MATRIX_STORED_LLS stored_lls, ENTRY *elist, int entry_count, int chunk_min)
+PAYLOADS build_matrix_get_payload_ladder(MATRIX_STORED_LLS stored_lls, ENTRY *elist, int32_t entry_count, int32_t chunk_min)
 {
     PAYLOADS payloads;
     payloads.arr = NULL;
     payloads.count = 0;
-    for (int i = 0; i < stored_lls.count; i++)
+    for (int32_t i = 0; i < stored_lls.count; i++)
     {
         MATRIX_STORED_LL curr_ll = stored_lls.stored_lls[i];
         PAYLOAD payload = deprecate_build_get_payload(elist, entry_count, curr_ll.full_load, curr_ll.zone, chunk_min);
@@ -381,27 +381,27 @@ PAYLOADS build_matrix_get_payload_ladder(MATRIX_STORED_LLS stored_lls, ENTRY *el
 /** \brief
  *
  * \param elist ENTRY*
- * \param entry_count int
- * \param chunk_border_sounds int
- * \param chunk_count int*
- * \param config int*
+ * \param entry_count int32_t
+ * \param chunk_border_sounds int32_t
+ * \param chunk_count int32_t*
+ * \param config int32_t*
  * \param permaloaded LIST
  *
  */
-void build_matrix_merge_random_main(ENTRY *elist, int entry_count, int chunk_border_sounds, int *chunk_count, int *config, LIST permaloaded)
+void build_matrix_merge_random_main(ENTRY *elist, int32_t entry_count, int32_t chunk_border_sounds, int32_t *chunk_count, int32_t *config, LIST permaloaded)
 {
 
     char temp[6] = "";
     char temp2[6] = "";
 
     double mult;
-    int iter_count, seed, max_payload_limit;
+    int32_t iter_count, seed, max_payload_limit;
     ask_params_matrix(&mult, &iter_count, &seed, &max_payload_limit);
     srand(seed);
 
     // for keeping track of the best found
-    long long int best_max = 9223372036854775807; // max signed 64b int
-    unsigned int best_zone = 0;
+    int64_t best_max = 9223372036854775807; // max signed 64b int32_t
+    uint32_t best_zone = 0;
 
     // clone elists that store the current iteration and the best iretation
     ENTRY clone_elist[ELIST_DEFAULT_SIZE];
@@ -413,36 +413,36 @@ void build_matrix_merge_random_main(ENTRY *elist, int entry_count, int chunk_bor
     build_assign_primary_chunks_all(elist, entry_count, chunk_count);                           // chunks start off having one entry per chunk
 
     // get occurence matrix and convert to the relation array (not doing this each iteration makes it way faster)
-    int permaloaded_include_flag = config[CNFG_IDX_MTRX_PERMA_INC_FLAG];
+    int32_t permaloaded_include_flag = config[CNFG_IDX_MTRX_PERMA_INC_FLAG];
     LIST entries = build_get_normal_entry_list(elist, entry_count);
     if (permaloaded_include_flag == 0)
-        for (int i = 0; i < permaloaded.count; i++)
+        for (int32_t i = 0; i < permaloaded.count; i++)
             list_remove(&entries, permaloaded.eids[i]);
 
-    int **entry_matrix = build_get_occurence_matrix(elist, entry_count, entries, config);
+    int32_t **entry_matrix = build_get_occurence_matrix(elist, entry_count, entries, config);
 
     // put the matrix's contents in an array and sort (so the matrix doesnt have to be searched every time)
     RELATIONS array_repr_untouched = build_transform_matrix(entries, entry_matrix, config, elist, entry_count);
     RELATIONS array_representation = build_transform_matrix(entries, entry_matrix, config, elist, entry_count);
     // matrix no longer used
-    for (int i = 0; i < entries.count; i++)
+    for (int32_t i = 0; i < entries.count; i++)
         free(entry_matrix[i]);
     free(entry_matrix);
 
     MATRIX_STORED_LLS stored_lls = build_matrix_store_lls(elist, entry_count);
 
     // runs until iter count is reached or until break inside goes off (iter count 0 can make it never stop)
-    for (int i = 0; i < iter_count || iter_count == 0; i++)
+    for (int32_t i = 0; i < iter_count || iter_count == 0; i++)
     {
         // copy elist into clone elist
-        for (int j = 0; j < entry_count; j++)
+        for (int32_t j = 0; j < entry_count; j++)
             clone_elist[j] = elist[j];
 
         // restore relations array so it can be tweaked again
         memcpy(array_representation.relations, array_repr_untouched.relations, array_representation.count * sizeof(RELATION));
 
         // second half of the matrix merge slightly randomised and ran on clone elist
-        for (int j = 0; j < array_representation.count; j++)
+        for (int32_t j = 0; j < array_representation.count; j++)
             array_representation.relations[j].value = array_repr_untouched.relations[j].value * randfrom(1.0, mult);
         qsort(array_representation.relations, array_representation.count, sizeof(RELATION), relations_cmp);
 
@@ -454,21 +454,21 @@ void build_matrix_merge_random_main(ENTRY *elist, int entry_count, int chunk_bor
         qsort(payloads.arr, payloads.count, sizeof(PAYLOAD), cmp_func_payload);
 
         // how many cam paths' payload it takes into consideration, max 8 but if the level is smaller it does less, also it could do 9 too but 8 is a nicer number
-        int src_depth = min(payloads.count, 8);
+        int32_t src_depth = min(payloads.count, 8);
 
         // gets current 'score', its multiplied by 100 instead of being bit shifted each iter so its legible in the console
-        long long int curr = 0;
-        for (int j = 0; j < src_depth; j++)
+        int64_t curr = 0;
+        for (int32_t j = 0; j < src_depth; j++)
         {
-            curr = curr * (long long int)100;
-            curr += (long long int)payloads.arr[j].count;
+            curr = curr * (int64_t)100;
+            curr += (int64_t)payloads.arr[j].count;
         }
 
-        int is_new_best = 0;
+        int32_t is_new_best = 0;
         // if its better than the previous best it gets stored in best_elist and score and zone are remembered
         if (curr < best_max)
         {
-            for (int j = 0; j < entry_count; j++)
+            for (int32_t j = 0; j < entry_count; j++)
                 best_elist[j] = clone_elist[j];
             best_max = curr;
             best_zone = payloads.arr[0].zone;
@@ -486,14 +486,14 @@ void build_matrix_merge_random_main(ENTRY *elist, int entry_count, int chunk_bor
     }
 
     // copy best_elist into the real elist used by build_main, correct chunk_count
-    for (int i = 0; i < entry_count; i++)
+    for (int32_t i = 0; i < entry_count; i++)
     {
         elist[i] = best_elist[i];
         if (elist[i].chunk >= *chunk_count)
             *chunk_count = elist[i].chunk + 1;
     }
 
-    for (int i = 0; i < stored_lls.count; i++)
+    for (int32_t i = 0; i < stored_lls.count; i++)
         free(stored_lls.stored_lls[i].full_load.eids);
 
     free(stored_lls.stored_lls);
