+++ src/build_merge_state_search.h
@@ -1,6 +1,6 @@
 // heap and hash stuff, dont want it in the main heap thing
 
-int heap_parent_index(int index)
+int32_t heap_parent_index(int32_t index)
 {
     return (index - 1) / 2;
 }
@@ -19,14 +19,14 @@ STATE_SEARCH_HEAP *heap_init_heap()
 // frees all dynamically allocated memory of the heap's contents and the heap
 void heap_destroy(STATE_SEARCH_HEAP *heap)
 {
-    for (unsigned int i = 0; i < heap->length; i++)
+    for (uint32_t i = 0; i < heap->length; i++)
         build_state_search_str_destroy(heap->heap_array[i]);
     free(heap->heap_array);
     free(heap);
 }
 
 // returns 1 if heap is empty, else 0
-int heap_is_empty(STATE_SEARCH_HEAP heap)
+int32_t heap_is_empty(STATE_SEARCH_HEAP heap)
 {
     if (heap.length == 0)
         return 1;
@@ -34,18 +34,18 @@ int heap_is_empty(STATE_SEARCH_HEAP heap)
 }
 
 // figures out what swap has to be done in heap when removing, 0 if none, 1 if element<->childL, 2 if element<->childR
-int heap_comp_util(STATE_SEARCH_HEAP *heap, int index_parent)
+int32_t heap_comp_util(STATE_SEARCH_HEAP *heap, int32_t index_parent)
 {
-    int index_child1 = 2 * index_parent + 1;
-    int index_child2 = 2 * index_parent + 2;
+    int32_t index_child1 = 2 * index_parent + 1;
+    int32_t index_child2 = 2 * index_parent + 2;
 
-    int temp = 0;
-    int length = heap->length;
+    int32_t temp = 0;
+    int32_t length = heap->length;
 
     if (index_child1 < length && index_child2 < length)
     {
-        unsigned int value1 = heap->heap_array[index_child1]->estimated + heap->heap_array[index_child1]->elapsed;
-        unsigned int value2 = heap->heap_array[index_child2]->estimated + heap->heap_array[index_child2]->elapsed;
+        uint32_t value1 = heap->heap_array[index_child1]->estimated + heap->heap_array[index_child1]->elapsed;
+        uint32_t value2 = heap->heap_array[index_child2]->estimated + heap->heap_array[index_child2]->elapsed;
         temp = (value1 < value2) ? 1 : 2;
     }
 
@@ -57,8 +57,8 @@ int heap_comp_util(STATE_SEARCH_HEAP *heap, int index_parent)
 
     if (temp == 1)
     {
-        unsigned int value1 = heap->heap_array[index_child1]->estimated + heap->heap_array[index_child1]->elapsed;
-        unsigned int valueP = heap->heap_array[index_parent]->estimated + heap->heap_array[index_parent]->elapsed;
+        uint32_t value1 = heap->heap_array[index_child1]->estimated + heap->heap_array[index_child1]->elapsed;
+        uint32_t valueP = heap->heap_array[index_parent]->estimated + heap->heap_array[index_parent]->elapsed;
         if (value1 < valueP)
             return 1;
         else
@@ -67,8 +67,8 @@ int heap_comp_util(STATE_SEARCH_HEAP *heap, int index_parent)
 
     if (temp == 2)
     {
-        unsigned int value2 = heap->heap_array[index_child2]->estimated + heap->heap_array[index_child2]->elapsed;
-        unsigned int valueP = heap->heap_array[index_parent]->estimated + heap->heap_array[index_parent]->elapsed;
+        uint32_t value2 = heap->heap_array[index_child2]->estimated + heap->heap_array[index_child2]->elapsed;
+        uint32_t valueP = heap->heap_array[index_parent]->estimated + heap->heap_array[index_parent]->elapsed;
         if (value2 < valueP)
             return 2;
         else
@@ -100,7 +100,7 @@ STATE_SEARCH_STR *heap_pop(STATE_SEARCH_HEAP *heap)
         return result;
     }
 
-    for (int n = 0, temp = heap_comp_util(heap, n); temp != 0; temp = heap_comp_util(heap, n))
+    for (int32_t n = 0, temp = heap_comp_util(heap, n); temp != 0; temp = heap_comp_util(heap, n))
     {
         STATE_SEARCH_STR *help = heap->heap_array[n];
         heap->heap_array[n] = heap->heap_array[2 * n + temp];
@@ -123,7 +123,7 @@ STATE_SEARCH_STR* heap_pop(STATE_SEARCH_HEAP *heap) {
 // adds state to the heap, makes necessary swap to mainatain the heap
 void heap_add(STATE_SEARCH_HEAP *heap, STATE_SEARCH_STR *input)
 {
-    int index_k = heap->length;
+    int32_t index_k = heap->length;
 
     heap->length += 1;
     if (heap->length >= heap->real_size)
@@ -143,9 +143,9 @@ void heap_add(STATE_SEARCH_HEAP *heap, STATE_SEARCH_STR *input)
         if (!index_k)
             break;
 
-        int index_p = heap_parent_index(index_k);
-        unsigned int value_k = heap->heap_array[index_k]->estimated + heap->heap_array[index_k]->elapsed;
-        unsigned int value_p = heap->heap_array[index_p]->estimated + heap->heap_array[index_p]->elapsed;
+        int32_t index_p = heap_parent_index(index_k);
+        uint32_t value_k = heap->heap_array[index_k]->estimated + heap->heap_array[index_k]->elapsed;
+        uint32_t value_p = heap->heap_array[index_p]->estimated + heap->heap_array[index_p]->elapsed;
 
         if (value_k >= value_p)
             break;
@@ -163,7 +163,7 @@ void heap_add(STATE_SEARCH_HEAP *heap, STATE_SEARCH_STR *input)
 */
 
 // initialises the table
-HASH_TABLE *hash_init_table(int (*hash_function)(HASH_TABLE *, unsigned short int *), int key_length)
+HASH_TABLE *hash_init_table(int32_t (*hash_function)(HASH_TABLE *, uint16_t *), int32_t key_length)
 {
     HASH_TABLE *table = (HASH_TABLE *)malloc(sizeof(HASH_TABLE)); // freed by caller
     table->table_length = HASH_TABLE_SIZE;
@@ -178,7 +178,7 @@ HASH_TABLE *hash_init_table(int (*hash_function)(HASH_TABLE *, unsigned short in
 void hash_destroy_table(HASH_TABLE *table)
 {
     HASH_ITEM *curr, *next;
-    for (int i = 0; i < table->table_length; i++)
+    for (int32_t i = 0; i < table->table_length; i++)
     {
         for (curr = table->items[i]; curr != NULL; curr = next)
         {
@@ -192,19 +192,19 @@ void hash_destroy_table(HASH_TABLE *table)
     free(table);
 }
 
-int hash_compare_keys(unsigned short int *array1, unsigned short int *array2, int arrays_length)
+int32_t hash_compare_keys(uint16_t *array1, uint16_t *array2, int32_t arrays_length)
 {
-    for (int i = 0; i < arrays_length; i++)
+    for (int32_t i = 0; i < arrays_length; i++)
         if (array1[i] != array2[i])
             return 0;
     return 1;
 }
 
-int hash_func_chek(HASH_TABLE *table, unsigned short int *entry_chunk_array)
+int32_t hash_func_chek(HASH_TABLE *table, uint16_t *entry_chunk_array)
 {
-    long long int hash_temp = 0;
-    int key_length = table->key_length;
-    for (int i = 0; i < key_length; i++)
+    int64_t hash_temp = 0;
+    int32_t key_length = table->key_length;
+    for (int32_t i = 0; i < key_length; i++)
     {
         hash_temp *= i + 1;
         hash_temp += entry_chunk_array[i];
@@ -215,18 +215,18 @@ int hash_func_chek(HASH_TABLE *table, unsigned short int *entry_chunk_array)
     if (hash_temp < 0)
         hash_temp += table->table_length;
 
-    return (int)hash_temp;
+    return (int32_t)hash_temp;
 }
 
-int hash_func(HASH_TABLE *table, unsigned short int *entry_chunk_array)
+int32_t hash_func(HASH_TABLE *table, uint16_t *entry_chunk_array)
 {
 
-    int key_length = table->key_length;
+    int32_t key_length = table->key_length;
 
-    int hash_temp = 0;
-    for (int i = 0; i < key_length; i++)
+    int32_t hash_temp = 0;
+    for (int32_t i = 0; i < key_length; i++)
     {
-        hash_temp ^= ((int)entry_chunk_array[i] << (4 * i % 27));
+        hash_temp ^= ((int32_t)entry_chunk_array[i] << (4 * i % 27));
     }
 
     hash_temp %= table->table_length;
@@ -236,11 +236,11 @@ int hash_func(HASH_TABLE *table, unsigned short int *entry_chunk_array)
     return hash_temp;
 }
 
-HASH_ITEM *hash_find(HASH_TABLE *table, unsigned short int *entry_chunk_array)
+HASH_ITEM *hash_find(HASH_TABLE *table, uint16_t *entry_chunk_array)
 // search function, returns valid pointer to the item or NULL if it does not exist
 {
     HASH_ITEM *curr;
-    int hash_value = table->hash_function(table, entry_chunk_array);
+    int32_t hash_value = table->hash_function(table, entry_chunk_array);
 
     if (table->items != NULL)
         for (curr = (table->items)[hash_value]; curr != NULL; curr = curr->next)
@@ -251,19 +251,19 @@ HASH_ITEM *hash_find(HASH_TABLE *table, unsigned short int *entry_chunk_array)
 }
 
 // initialises hash item
-HASH_ITEM *hash_make_item(unsigned short int *entry_chunk_array, int key_length)
+HASH_ITEM *hash_make_item(uint16_t *entry_chunk_array, int32_t key_length)
 {
     HASH_ITEM *item = NULL;
     item = (HASH_ITEM *)malloc(sizeof(HASH_ITEM)); // freed when table is destroyed
-    item->entry_chunk_array = (unsigned short int *)malloc(key_length * sizeof(unsigned short int));
-    memcpy(item->entry_chunk_array, entry_chunk_array, key_length * sizeof(unsigned short int));
+    item->entry_chunk_array = (uint16_t *)malloc(key_length * sizeof(uint16_t));
+    memcpy(item->entry_chunk_array, entry_chunk_array, key_length * sizeof(uint16_t));
     item->next = NULL;
 
     return item;
 }
 
 // attempts to add an item into the specified list of the table, if such an item already exists returns 0, else returns 1 and adds
-int hash_try_to_add_item(HASH_TABLE *table, HASH_ITEM *item, int list_index)
+int32_t hash_try_to_add_item(HASH_TABLE *table, HASH_ITEM *item, int32_t list_index)
 {
     HASH_ITEM *curr, *last = NULL;
 
@@ -288,16 +288,16 @@ int hash_try_to_add_item(HASH_TABLE *table, HASH_ITEM *item, int list_index)
 }
 
 // puts the item in the specified list (used in resize)
-void hash_place_item(HASH_TABLE *table, HASH_ITEM *item, int list_index)
+void hash_place_item(HASH_TABLE *table, HASH_ITEM *item, int32_t list_index)
 {
     item->next = (table->items)[list_index];
     (table->items)[list_index] = item;
 }
 
 // attempts to add an item with this key, 'fails' if such an item already exists
-int hash_add(HASH_TABLE *table, unsigned short int *entry_chunk_array)
+int32_t hash_add(HASH_TABLE *table, uint16_t *entry_chunk_array)
 {
-    int hash_val = table->hash_function(table, entry_chunk_array);
+    int32_t hash_val = table->hash_function(table, entry_chunk_array);
 
     // if adding it was not successful it gets rid of the item and returns, else goes on
     HASH_ITEM *curr = hash_make_item(entry_chunk_array, table->key_length);
@@ -312,16 +312,16 @@ int hash_add(HASH_TABLE *table, unsigned short int *entry_chunk_array)
     table->item_count += 1;
     if (((double) table->item_count / table->table_length) >= HASH_FULLNESS_RATIO) {
         HASH_ITEM** things = (HASH_ITEM**) malloc((table->item_count + 1) * sizeof(HASH_ITEM *));       // unused
-        int index = 0;
+        int32_t index = 0;
         if (things != NULL)
         {
-            for (int i = 0; i < table->table_length; i++)
+            for (int32_t i = 0; i < table->table_length; i++)
                 for (curr = (table->items)[i]; curr != NULL; curr = curr->next)
                     things[index++] = curr;
             free(table->items);
             table->items = (HASH_ITEM**) calloc(HASH_RESIZE_MULT * table->table_length, sizeof(HASH_ITEM*)); // unused
             table->table_length = table->table_length * HASH_RESIZE_MULT;
-            for (int i = 0; i < index; i++)
+            for (int32_t i = 0; i < index; i++)
             {
                 things[i]->next = NULL;
                 hash_place_item(table, things[i], table->hash_function(things[i]->entry_chunk_array, table->key_length, table->table_length));
